Refactor folder icons to allow them to be used with other buttons
