+++ src/app/Components/ControlPanel/ButtonIcon.tsx
@@ -1,5 +1,6 @@
-import React from 'react'
+import React, { forwardRef } from 'react'
 import { IconType } from '@react-icons/all-files/lib'
+import { Group, SelectItem as MantineSelectItem, Avatar, Text } from '@mantine/core'
 
 import { FaVideo } from '@react-icons/all-files/fa/FaVideo'
 import { FaLightbulb } from '@react-icons/all-files/fa/FaLightbulb'
@@ -24,9 +25,10 @@ export const AvailableIcons = {
 	FaMicrophone: 'Microphone',
 	FaMicrophoneAlt: 'Microphone',
 	FaVideo: 'Camera',
+	FaFolder: 'Folder',
 }
 
-export const FolderIcon = (icon: string): IconType => {
+export const ButtonIcon = (icon: string): IconType => {
 	if (icon === 'FaLightbulb') return FaLightbulb
 	else if (icon === 'FaRegLightbulb') return FaRegLightbulb
 	else if (icon === 'FaVolumeDown') return FaVolumeDown
@@ -37,10 +39,11 @@ export const FolderIcon = (icon: string): IconType => {
 	else if (icon === 'FaMicrophone') return FaMicrophone
 	else if (icon === 'FaMicrophoneAlt') return FaMicrophoneAlt
 	else if (icon === 'FaVideo') return FaVideo
-	else return FaFolder
+	else if (icon === 'FaFolder') return FaFolder
+	else return null
 }
 // TODO surely we don't need two functions?
-export const FolderIconReact = (props: { icon: string }) => {
+export const ButtonIconReact = (props: { icon: string }) => {
 	if (props.icon === 'FaLightbulb') return <FaLightbulb />
 	else if (props.icon === 'FaRegLightbulb') return <FaRegLightbulb />
 	else if (props.icon === 'FaVolumeDown') return <FaVolumeDown />
@@ -51,5 +54,27 @@ export const FolderIconReact = (props: { icon: string }) => {
 	else if (props.icon === 'FaMicrophone') return <FaMicrophone />
 	else if (props.icon === 'FaMicrophoneAlt') return <FaMicrophoneAlt />
 	else if (props.icon === 'FaVideo') return <FaVideo />
-	else return <FaFolder />
+	else if (props.icon === 'FaFolder') return <FaFolder />
+	else return <></>
 }
+
+interface ItemProps extends React.ComponentPropsWithoutRef<'div'> {
+	icon: string
+	label: string
+}
+
+// eslint-disable-next-line react/display-name
+export const ButtonIconSelectItem = forwardRef<HTMLDivElement, ItemProps>(
+	({ icon, label, ...others }: ItemProps, ref) => (
+		<div ref={ref} {...others}>
+			<Group noWrap>
+				<Avatar radius={'xs'} size={'md'}>
+					<ButtonIconReact icon={icon} />
+				</Avatar>
+				<div>
+					<Text size="sm">{label}</Text>
+				</div>
+			</Group>
+		</div>
+	)
+)
